reorganize controlplane integration tests
