+++ test/integration/controlplane/transformation/kms_transformation_test.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controlplane
+package transformation
 
 import (
 	"bytes"
